delete row alert

+++ frontend/app/novels/novels-list/delete-row-button.tsx
@@ -1,24 +1,55 @@
 import * as React from "react"
 
 import { Button } from "../../../components/ui/button"
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@/components/ui/alert-dialog"
+
 import { fetch_backend } from "@/utils/fetch_backend"
-import { NovelEntry } from "./novel-types"
+import { NovelEntry } from "./novel-types.ts"
 
 interface DeleteRowButtonProps {
   row: any,
   table: any,
 }
 
-export function DeleteRowButton({ row, table} : DeleteRowButtonProps) {
+export function DeleteRowButton({ row, table } : DeleteRowButtonProps) {
   return (
-    <Button
-      variant="secondary"
-      size={"sm"}
-      onClick={() => delete_row({row, table})}
-      className='text-red-500'
-    >
-      Delete Row
-    </Button>
+    <AlertDialog>
+      <AlertDialogTrigger asChild>
+        <Button 
+          variant="secondary"
+          size={"sm"}
+          className='text-red-500'
+        >
+          Delete Row
+        </Button>
+      </AlertDialogTrigger>
+      <AlertDialogContent>
+        <AlertDialogHeader>
+          <AlertDialogTitle>Are you sure? This will delete the following row:</AlertDialogTitle>
+          <AlertDialogDescription>
+            {novel_to_component(row.original)}
+          </AlertDialogDescription>
+        </AlertDialogHeader>
+        <AlertDialogFooter>
+          <AlertDialogCancel>Cancel</AlertDialogCancel>
+          <AlertDialogAction
+            onClick={() => delete_row({row, table})}
+          >
+            Continue
+          </AlertDialogAction>
+        </AlertDialogFooter>
+      </AlertDialogContent>
+    </AlertDialog>
   )
 }
 
@@ -39,4 +70,17 @@ function delete_row({ row, table } : DeleteRowButtonProps) {
   }
 
   table.options.meta.updateTableData(data);
+}
+
+// parses a novel entry as a minimized component
+function novel_to_component(novel: NovelEntry) {
+  return (
+    <div className="pt-2 text-primary">
+      <div className="text-sm">Title: {novel.title}</div>
+      <div className="text-sm">Chapter: {novel.chapter}</div>
+      <div className="text-sm">Rating: {novel.rating}</div>
+      <div className="text-sm">Status: {novel.status}</div>
+      <div className="text-sm">Last Updated: {novel.date_modified.toString()}</div>
+    </div>
+  )
 }
